+++ src/Digdir.Domain.Dialogporten.GraphQL/EndUser/SearchDialogs/ObjectTypes.cs
@@ -1,7 +1,3 @@
-using System.Globalization;
-using System.Text;
-using Digdir.Domain.Dialogporten.Application.Common.Pagination.Order;
-using Digdir.Domain.Dialogporten.Application.Features.V1.EndUser.Dialogs.Queries.Search;
 using Digdir.Domain.Dialogporten.GraphQL.EndUser.Common;
 using Digdir.Domain.Dialogporten.GraphQL.EndUser.MutationTypes;
 
@@ -37,7 +33,7 @@ public sealed class SearchDialogsPayload
     public string? ContinuationToken { get; set; }
 
     [GraphQLDescription("Use this OrderBy to fetch the next page of dialogs, must be used in combination with ContinuationToken")]
-    public List<SearchDialogSortType> OrderBy { get; set; } = null!;
+    public List<SearchDialogSortType> OrderBy { get; set; } = [];
 
     public List<ISearchDialogError> Errors { get; set; } = [];
 }
@@ -142,77 +138,4 @@ public sealed class SearchDialogInput
 
     [GraphQLDescription("Sort the results by one or more fields")]
     public List<SearchDialogSortType>? OrderBy { get; set; }
-}
-
-internal static class SearchDialogSortTypeExtensions
-{
-    public static List<SearchDialogSortType> ToSearchDialogSortTypeList(
-        this string orderBy)
-    {
-        List<SearchDialogSortType> searchDialogSortTypes = [];
-
-        orderBy = orderBy
-            .ToLower(CultureInfo.InvariantCulture)
-            .Replace("id_desc", "", StringComparison.OrdinalIgnoreCase)
-            .Replace("id_asc", "", StringComparison.OrdinalIgnoreCase);
-
-        foreach (var orderByPart in orderBy.Split(','))
-        {
-            var parts = orderByPart.Split('_');
-            if (parts.Length != 2)
-            {
-                continue;
-            }
-
-            var sortDirection = parts[1] switch
-            {
-                "asc" => SortDirection.Asc,
-                "desc" => SortDirection.Desc,
-                _ => throw new InvalidOperationException("Invalid sort direction")
-            };
-
-            searchDialogSortTypes.Add(parts[0] switch
-            {
-                "createdat" => new SearchDialogSortType { CreatedAt = sortDirection },
-                "updatedat" => new SearchDialogSortType { UpdatedAt = sortDirection },
-                "dueat" => new SearchDialogSortType { DueAt = sortDirection },
-                _ => throw new InvalidOperationException("Invalid sort field")
-            });
-        }
-
-        return searchDialogSortTypes;
-    }
-
-    public static bool TryToOrderSet(this List<SearchDialogSortType> searchDialogSortTypes,
-        out OrderSet<SearchDialogQueryOrderDefinition, IntermediateDialogDto>? orderSet)
-    {
-        var stringBuilder = new StringBuilder();
-        foreach (var orderBy in searchDialogSortTypes)
-        {
-            if (orderBy.CreatedAt != null)
-            {
-                stringBuilder.Append(CultureInfo.InvariantCulture, $"createdAt_{orderBy.CreatedAt},");
-                continue;
-            }
-            if (orderBy.UpdatedAt != null)
-            {
-                stringBuilder.Append(CultureInfo.InvariantCulture, $"updatedAt_{orderBy.UpdatedAt},");
-                continue;
-            }
-            if (orderBy.DueAt != null)
-            {
-                stringBuilder.Append(CultureInfo.InvariantCulture, $"dueAt_{orderBy.DueAt},");
-            }
-        }
-
-        if (OrderSet<SearchDialogQueryOrderDefinition, IntermediateDialogDto>.TryParse(stringBuilder.ToString(),
-                out var parsedOrderSet))
-        {
-            orderSet = parsedOrderSet;
-            return true;
-        }
-
-        orderSet = null;
-        return false;
-    }
 }
